requirements/dev.txt
-r base.txt
ipython>=8.20.0
django-debug-toolbar>=4.2.0
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
coverage>=7.4.0
